chore: 타입분리
