+++ src/services/auth.service.ts
@@ -1,21 +1,7 @@
 import axiosApi from '@/helper/api_helper'
 import { apiUrl } from '@/shared/constants/configure'
+import { LoginRequest, LoginResponse, SignupRequest } from '@/shared/types/auth'
 import useSessionStore from '@/store/useSessionStore'
-
-interface LoginRequest {
-  email: string
-  password: string
-}
-
-interface SignupRequest {
-  email: string
-  password: string
-  nickname: string
-}
-
-interface LoginResponse {
-  token: string
-}
 
 export const postLogin = async (
   payload: LoginRequest,
@@ -24,16 +10,13 @@ export const postLogin = async (
   return response.data
 }
 
-export const postLogout = async () => {
+export const getLogout = async () => {
   console.log('로그아웃 시도')
-  console.log('현재 토큰:', useSessionStore.getState()?.access_token) // 토큰이 있는지 확인
+  console.log('현재 토큰:', useSessionStore.getState()?.access_token)
 
   try {
     const response = await axiosApi.get('/users/logout')
-    console.log('로그아웃 응답:', response)
 
-    document.cookie =
-      'refresh_token=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;'
     return response.data
   } catch (error) {
     console.log('로그아웃 에러:', error)
