+++ src/shared/types/auth.ts
@@ -1,0 +1,14 @@
+export interface LoginRequest {
+  email: string
+  password: string
+}
+
+export interface SignupRequest {
+  email: string
+  password: string
+  nickname: string
+}
+
+export interface LoginResponse {
+  token: string
+}
